Updates clap and makes it optional.
For library users pulling in clap might be undesired, as it's only used
in `src/main.rs` I made it optional (but default to not impact binary
installs).

This PR also updates clap and keeps using the clap builder API (so no
heavy syn/proc-macro dependencies) but updates to the latest clap which
removes some dependencies that will flag in audit runs.